Moved notification to vanilla folder
