+++ src/common/components/vanilla/notification/notification.css
